+++ R/ETFs.R
@@ -1,6 +1,6 @@
 #' Daily prices and total returns for 24 ETFs.
 #'
-#' Data set containing daily prices and total returns for 24 exchange-traded
+#' Data set containing daily prices and total returns for 37 exchange-traded
 #' funds (ETFs) as well as daily returns for U.S. Treasury bills
 #' (risk-free asset).
 #'