+++ R/tactical_DualMomentum.R
@@ -2,7 +2,7 @@
 #'
 #' \code{tactical_DualMomentum} determines asset allocations for a strategy
 #' according to the dual momentum approach described in Antonacci (2016)
-#' <doi:https://dx.doi.org/10.2139/ssrn.2042750>.
+#' <https://dx.doi.org/10.2139/ssrn.2042750>.
 #'
 #' Dual momentum sorts assets within each asset class described in \code{strat}
 #' on a relative basis (i.e. which asset outperforms others within the same