+++ R/tactical_TrendFriend.R
@@ -18,9 +18,13 @@
 #' @param R An xts object with daily returns of the tickers in strat.
 #' @param risk_free Either an xts object with daily returns of the risk-free
 #' @examples
-#' ivy  <- asset_allocations$tactical$ivy
+#' trend_friend  <- asset_allocations$tactical$trend_friend
 #' reb_date <- as.Date("2022-03-31")
-#' tactical_TrendFriend(ivy, reb_date, ETFs$Prices[, ivy$tickers], ETFs$Returns[, ivy$tickers])
+#' tactical_TrendFriend(trend_friend,
+#'                      reb_date,
+#'                      ETFs$Prices[, trend_friend$tickers],
+#'                      ETFs$Returns[, trend_friend$tickers]
+#'                     )
 #' @return A numeric vector of weights after applying the rule.
 #' @export
 #' @import xts
@@ -72,7 +76,7 @@ tactical_TrendFriend <- function(strat, reb_date, P, R, risk_free = NULL){
   # step 1: calculation of weights using inverse-volatility weights
   if (nrow(P_month_ends) >= max(n_months_trend, n_months_vol)){
     w <- 1/apply(R[seq(from = nrow(R) - 21*n_months_vol +1,
-                       to = nrow(R)), ], 2, sd)
+                       to = nrow(R)), ], 2, stats::sd)
     w <- w/sum(w)
 
     # step 2: trend following rule