+++ data-raw/DATASET.R
@@ -1,3 +1,4 @@
 ## code to prepare `DATASET` dataset goes here
 
 usethis::use_data(DATASET, overwrite = TRUE)
+